Hack together window creation with SDL3

+++ src/InitWindow.cpp
@@ -232,7 +232,7 @@ RGAPI void* InitWindow(uint32_t width, uint32_t height, const char* title){
     if(!(g_renderstate.windowFlags & FLAG_HEADLESS)){
         #if SUPPORT_GLFW == 1 || SUPPORT_SDL2 == 1 || SUPPORT_SDL3 == 1 || SUPPORT_RGFW == 1
         #ifdef MAIN_WINDOW_GLFW
-        SubWindow createdWindow = InitWindow_GLFW(width, height, title);
+        SubWindow createdWindow = InitWindow_GLFW((int)width, (int)height, title);
         #elif defined(MAIN_WINDOW_SDL3)
         Initialize_SDL3();
         SubWindow createdWindow = InitWindow_SDL3(width, height, title);
